+++ segmodels_keras/models/fpn.py
@@ -108,6 +108,7 @@ def build_fpn(
     aggregation="sum",
     dropout=None,
     weights_notop=None,
+    freeze_notop=False,
 ):
     input_ = backbone.input
     x = backbone.output
@@ -177,6 +178,8 @@ def build_fpn(
     if weights_notop is not None:
         model_notop = models.Model(input_, x)
         model_notop.load_weights(weights_notop)
+        if freeze_notop:
+            freeze_model(model_notop)
         x = model_notop.output
 
     # model head (define number of output classes)
@@ -208,6 +211,7 @@ def FPN(
     activation="softmax",
     weights=None,
     weights_notop=None,
+    freeze_notop=False,
     encoder_weights="imagenet",
     encoder_freeze=False,
     encoder_features="default",
@@ -230,6 +234,8 @@ def FPN(
         weights: optional, path to model weights to be loaded.
         weights_notop: optional, path to model weights without top (without segmentation
             head) to be loaded.
+        freeze_notop: if ``True``, set all layers of the model except the top as
+            non-trainable.
         activation: name of one of ``keras.activations`` for last model layer
             (e.g. ``sigmoid``, ``softmax``, ``linear``).
         encoder_weights: one of ``None`` (random initialization), ``imagenet``
@@ -276,6 +282,7 @@ def FPN(
         classes=classes,
         aggregation=pyramid_aggregation,
         weights_notop=weights_notop,
+        freeze_notop=freeze_notop,
     )
 
     # lock encoder weights for fine-tuning