+++ segmodels_keras/models/unet.py
@@ -114,6 +114,7 @@ def build_unet(
     activation="sigmoid",
     use_batchnorm=True,
     weights_notop=None,
+    freeze_notop=False,
 ):
     input_ = backbone.input
     x = backbone.output
@@ -146,6 +147,8 @@ def build_unet(
     if weights_notop is not None:
         model_notop = models.Model(input_, x)
         model_notop.load_weights(weights_notop)
+        if freeze_notop:
+            freeze_model(model_notop)
         x = model_notop.output
 
     # model head (define number of output classes)
@@ -177,6 +180,7 @@ def Unet(
     activation="sigmoid",
     weights=None,
     weights_notop=None,
+    freeze_notop=False,
     encoder_weights="imagenet",
     encoder_freeze=False,
     encoder_features="default",
@@ -201,6 +205,8 @@ def Unet(
         weights: optional, path to model weights to be loaded.
         weights_notop: optional, path to model weights without top (without segmentation
             head) to be loaded.
+        freeze_notop: if ``True``, set all layers of the model except the top as
+            non-trainable.
         encoder_weights: one of ``None`` (random initialization), ``imagenet``
             (pre-training on ImageNet).
         encoder_freeze: if ``True`` set all layers of encoder (backbone model) as
@@ -256,6 +262,7 @@ def Unet(
         n_upsample_blocks=len(decoder_filters),
         use_batchnorm=decoder_use_batchnorm,
         weights_notop=weights_notop,
+        freeze_notop=freeze_notop,
     )
 
     # lock encoder weights for fine-tuning